// bedrock-types.ts
export type ShortString = string;
/**
 * UUID is the UUID of the command called. This UUID is a bit odd as it is not specified by the server. It
 * is not clear what exactly this UUID is meant to identify, but it is unique for each command called.
 */
export type uuid = string;
export type byterot = number;
export type restBuffer = Buffer;
export type encapsulated = any;
export type nbt = any;
// ...
export type ItemExtraDataWithBlockingTick = {
  has_nbt: "false" | "true";
  nbt?: {
    version: number;
    nbt: any;
  };
  can_place_on: ShortString[];
  can_destroy: ShortString[];
  blocking_tick: bigint;
};
export type ItemExtraDataWithoutBlockingTick = {
  has_nbt: "false" | "true";
  nbt?: {
    version: number;
    nbt: any;
  };
  can_place_on: ShortString[];
  can_destroy: ShortString[];
};
/**
 * Same as below but without a "networkStackID" boolean
 */
export type ItemLegacy = {
  network_id: number;
  count?: number;
  metadata?: number;
  block_runtime_id?: number;
  extra?:
    | ItemExtraDataWithBlockingTick
    | ItemExtraDataWithoutBlockingTick
    | undefined;
};
// ...
export type RecipeIngredient = {
  type:
    | "invalid"
    | "int_id_meta"
    | "molang"
    | "item_tag"
    | "string_id_meta"
    | "complex_alias";
  network_id?: number;
  metadata?: number | undefined | number;
  expression?: string;
  version?: number;
  tag?: string;
  name?: string;
  count: number;
};
// ...
export type StackRequestSlotInfo = {
  /** ContainerID is the ID of the container that the slot was in. */
  slot_type: FullContainerName;
  /** Slot is the index of the slot within the container with the ContainerID above. */
  slot: number;
  /** StackNetworkID is the unique stack ID that the client assumes to be present in this slot. The server must check if these IDs match. If they do not match, servers should reject the stack request that the action holding this info was in. */
  stack_id: number;
};
/**
 * ItemStackRequest is sent by the client to change item stacks in an inventory. It is essentially a
 * replacement of the InventoryTransaction packet added in 1.16 for inventory specific actions, such as moving
 * items around or crafting. The InventoryTransaction packet is still used for actions such as placing blocks
 * and interacting with entities.
 */
export type ItemStackRequest = {
  /** RequestID is a unique ID for the request. This ID is used by the server to send a response for this specific request in the ItemStackResponse packet. */
  request_id: number;
  actions: {
    type_id:
      | "take"
      | "place"
      | "swap"
      | "drop"
      | "destroy"
      | "consume"
      | "create"
      | "place_in_container"
      | "take_out_container"
      | "lab_table_combine"
      | "beacon_payment"
      | "mine_block"
      | "craft_recipe"
      | "craft_recipe_auto"
      | "craft_creative"
      | "optional"
      | "craft_grindstone_request"
      | "craft_loom_request"
      | "non_implemented"
      | "results_deprecated";
    count?: number;
    source?: StackRequestSlotInfo;
    destination?: StackRequestSlotInfo;
    randomly?: boolean;
    result_slot_id?: number;
    primary_effect?: number;
    secondary_effect?: number;
    hotbar_slot?: number;
    predicted_durability?: number;
    network_id?: number;
    recipe_network_id?: number;
    times_crafted?: number;
    times_crafted_2?: number;
    ingredients?: RecipeIngredient[];
    item_id?: number;
    filtered_string_index?: number;
    cost?: number;
    pattern?: string;
    result_items?: ItemLegacy[];
  }[];
  custom_names: string[];
  cause:
    | "chat_public"
    | "chat_whisper"
    | "sign_text"
    | "anvil_text"
    | "book_and_quill_text"
    | "command_block_text"
    | "block_actor_data_text"
    | "join_event_text"
    | "leave_event_text"
    | "slash_command_chat"
    | "cartography_text"
    | "kick_command"
    | "title_command"
    | "summon_command";
};
/**
 * ItemStackResponse is a response to an individual ItemStackRequest.
 */
export type ItemStackResponses = {
  /** Status specifies if the request with the RequestID below was successful. If this is the case, the ContainerInfo below will have information on what slots ended up changing. If not, the container info will be empty. A non-0 status means an error occurred and will result in the action being reverted. */
  status: "ok" | "error";
  /** RequestID is the unique ID of the request that this response is in reaction to. If rejected, the client will undo the actions from the request with this ID. */
  request_id: number;
  containers?: {
    slot_type: FullContainerName;
    slots: {
      slot: number;
      hotbar_slot: number;
      count: number;
      item_stack_id: number;
      custom_name: string;
      filtered_custom_name: string;
      durability_correction: number;
    }[];
  }[];
}[];
// ...
export type ContainerSlotType =
  | "anvil_input"
  | "anvil_material"
  | "anvil_result"
  | "smithing_table_input"
  | "smithing_table_material"
  | "smithing_table_result"
  | "armor"
  | "container"
  | "beacon_payment"
  | "brewing_input"
  | "brewing_result"
  | "brewing_fuel"
  | "hotbar_and_inventory"
  | "crafting_input"
  | "crafting_output"
  | "recipe_construction"
  | "recipe_nature"
  | "recipe_items"
  | "recipe_search"
  | "recipe_search_bar"
  | "recipe_equipment"
  | "recipe_book"
  | "enchanting_input"
  | "enchanting_lapis"
  | "furnace_fuel"
  | "furnace_ingredient"
  | "furnace_output"
  | "horse_equip"
  | "hotbar"
  | "inventory"
  | "shulker"
  | "trade_ingredient1"
  | "trade_ingredient2"
  | "trade_result"
  | "offhand"
  | "compcreate_input"
  | "compcreate_output"
  | "elemconstruct_output"
  | "matreduce_input"
  | "matreduce_output"
  | "labtable_input"
  | "loom_input"
  | "loom_dye"
  | "loom_material"
  | "loom_result"
  | "blast_furnace_ingredient"
  | "smoker_ingredient"
  | "trade2_ingredient1"
  | "trade2_ingredient2"
  | "trade2_result"
  | "grindstone_input"
  | "grindstone_additional"
  | "grindstone_result"
  | "stonecutter_input"
  | "stonecutter_result"
  | "cartography_input"
  | "cartography_additional"
  | "cartography_result"
  | "barrel"
  | "cursor"
  | "creative_output"
  | "smithing_table_template"
  | "crafter"
  | "dynamic"
  | "registry";
// ...
export type FullContainerName = {
  container_id: ContainerSlotType;
  dynamic_container_id?: number;
};
// ...
export type packet_item_stack_response = {
  /** Responses is a list of responses to ItemStackRequests sent by the client before. Responses either approve or reject a request from the client. Vanilla limits the size of this slice to 4096. */
  responses: ItemStackResponses;
};